Refactor data-entity
