+++ packages/utils/src/data-entity/utils.ts
@@ -1,0 +1,9 @@
+export function setDataEntityProps(input: any): void {
+    Object.defineProperties(input, {
+        __isDataEntity: {
+            value: true,
+            enumerable: false,
+            writable: false,
+        }
+    });
+}
