+++ packages/utils/src/index.ts
@@ -4,6 +4,7 @@ export * from './arrays';
 export * from './big-map';
 export * from './collector';
 export * from './data-entity';
+export * from './data-entity/interfaces';
 export * from './dates';
 export * from './interfaces';
 export * from './logger';
